In Memory Vote data
Create the groundwork for adding nominations to the memory repo.
logic contains:
adding new nomination.
validating that the nomination does not already exist.
creates the vote for the nomination.
verify that the user has not already existing vote for the Nomination.
if the vote for the nomination already exists for the user, allows for upvote.
Added a lot of the classes that will be required for the entire project.